+++ src/core/env/env.go
@@ -15,7 +15,7 @@ var (
 	// Mode - текущий режим работы системы.
 	Mode env_mode.Mode = env_mode.Dev
 	// Version - текущая версия системы.
-	Version = "24.0.7"
+	Version = "24.0.8"
 	// OS - версия операционной системы.
 	OS = fmt.Sprintf("%s - %s", runtime.GOOS, runtime.GOARCH)
 )