+++ src/core/internal_test.go
@@ -5,7 +5,11 @@ import (
 )
 
 func Test_core_Ctx(t *testing.T) {
-	var cr, _ = New()
+	var cr, err = New()
+
+	if err != nil {
+		t.Fatalf("Failed to create the system core: '%s'. ", err)
+	}
 
 	if ctx := cr.Ctx(); ctx == nil {
 		t.Error("Context are nil. ")
@@ -13,29 +17,31 @@ func Test_core_Ctx(t *testing.T) {
 }
 
 func Test_core_Components(t *testing.T) {
-	var (
-		cr, _      = New()
-		components = cr.Components()
-	)
+	var cr, err = New()
+
+	if err != nil {
+		t.Fatalf("Failed to create the system core: '%s'. ", err)
+	}
 
 	switch {
-	case components == nil:
+	case cr.Components() == nil:
 		t.Error("Components is nil. ")
-	case components.Logger() == nil:
+	case cr.Components().Logger() == nil:
 		t.Error("Logger component is nil. ")
 	}
 }
 
 func Test_core_Tools(t *testing.T) {
-	var (
-		cr, _ = New()
-		tools = cr.Tools()
-	)
+	var cr, err = New()
+
+	if err != nil {
+		t.Fatalf("Failed to create the system core: '%s'. ", err)
+	}
 
 	switch {
-	case tools == nil:
+	case cr.Tools() == nil:
 		t.Error("Tools is nil. ")
-	case tools.TaskScheduler() == nil:
+	case cr.Tools().TaskScheduler() == nil:
 		t.Error("Task scheduler tool is nil. ")
 	}
 }