+++ src/core/tools.go
@@ -5,6 +5,11 @@ import (
 	"sm-box/src/core/tools/task_scheduler"
 )
 
+// Tools - описание внутренних инструментов ядра системы.
+type Tools interface {
+	TaskScheduler() task_scheduler.Scheduler
+}
+
 // tools - внутренние инструменты ядра системы.
 type tools struct {
 	closer        closer.Closer