+++ src/core/tools/closer/closer.go
@@ -3,7 +3,7 @@ package closer
 import (
 	"context"
 	"sm-box/src/core/components/tracer"
-	"sync"
+	"sm-box/src/core/env"
 )
 
 // Closer - описание инструмента ядра системы отвечающий за корректное завершение работы системы.
@@ -13,25 +13,33 @@ type Closer interface {
 }
 
 // New - создание инструмента ядра системы отвечающий за корректное завершение работы системы.
-func New(conf *Config, ctx context.Context, wg *sync.WaitGroup) (cl Closer, ct context.Context) {
+func New(conf *Config, ctx context.Context) (cl Closer, ct context.Context) {
 	// tracer
 	{
 		var trc = tracer.New(tracer.LevelMain, tracer.LevelCoreTool)
 
-		trc.FunctionCall(conf, ctx, wg)
+		trc.FunctionCall(conf, ctx)
 		trc.FunctionCallFinished(cl, ct)
 	}
 
 	var c = &closer{
 		conf: conf,
 
-		wg: wg,
+		stop: make(chan struct{}, 5),
 	}
 
 	c.ctx, c.ctxCancel = context.WithCancel(ctx)
 
 	cl = c
 	ct = c.ctx
+
+	env.Synchronization.WaitGroup.Add(1)
+
+	go func() {
+		defer env.Synchronization.WaitGroup.Done()
+
+		c.tracking()
+	}()
 
 	return
 }