+++ src/internal/server/controllers.go
@@ -1,0 +1,8 @@
+package server
+
+// Controllers - описание контроллеров сервера.
+type Controllers interface {
+}
+
+// controllers - контроллеры сервера.
+type controllers struct{}